carousel buttons are not overlapping cards anymore. Better spacing to carousel cards. Highlighting is disabled now
