+++ client/src/components/BudgetCardList/BudgetCardList.styles.ts
@@ -4,10 +4,12 @@ import { motion } from "framer-motion";
 export const Container = styled.div`
   position: relative;
   width: 100%;
-  height: 80%;
+  height: 70%;
   overflow: hidden;
   display: flex;
   justify-content: center;
+  user-select: none;
+  margin-top: 2em;
 `;
 
 export const Carousel = styled(motion.div)`
@@ -16,30 +18,6 @@ export const Carousel = styled(motion.div)`
   justify-content: center;
   height: 100vh;
   width: 100%;
-`;
-
-export const UpArrow = styled.div`
-  position: absolute;
-  top: 0;
-  z-index: 1;
-  cursor: pointer;
-  transition: 0.3s all ease;
-
-  &:hover {
-    transform: scale(1.1);
-  }
-`;
-
-export const DownArrow = styled.div`
-  position: absolute;
-  bottom: 0;
-  z-index: 1;
-  cursor: pointer;
-  transition: 0.3s all ease;
-
-  &:hover {
-    transform: scale(1.1);
-  }
 `;
 
 export const GotoCurrent = styled.div`
