+++ client/src/components/BudgetCardList/BudgetCardList.tsx
@@ -1,36 +1,25 @@
 import React, { useEffect, useState } from "react";
-import {
-  Container,
-  Carousel,
-  UpArrow,
-  DownArrow,
-  GotoCurrent,
-} from "./BudgetCardList.styles";
+import { Container, Carousel, GotoCurrent } from "./BudgetCardList.styles";
 import { BudgetCard } from "../../components";
-import { IoIosArrowUp, IoIosArrowDown } from "react-icons/io";
+
 import { IBudgetData } from "../../constants";
 
 interface IBudgetCardList {
   budgets: IBudgetData[];
   refetch: () => void;
+  position: number;
+  setPosition: React.Dispatch<React.SetStateAction<number>>;
 }
 
-const BudgetCardList: React.FC<IBudgetCardList> = ({ budgets, refetch }) => {
-  const [position, setPosition] = useState<number>(12);
+const BudgetCardList: React.FC<IBudgetCardList> = ({
+  budgets,
+  refetch,
+  position,
+  setPosition,
+}) => {
   const [rerender, setRerender] = useState<boolean>(true);
 
   const toggleRerender = () => setRerender(!rerender);
-
-  const onDown = () => {
-    if (position < budgets.length - 1) {
-      setPosition(position + 1);
-    }
-  };
-  const onUp = () => {
-    if (position > 0) {
-      setPosition(position - 1);
-    }
-  };
 
   useEffect(() => {
     refetch();
@@ -38,10 +27,6 @@ const BudgetCardList: React.FC<IBudgetCardList> = ({ budgets, refetch }) => {
 
   return (
     <Container>
-      <UpArrow onClick={onUp}>
-        <IoIosArrowUp size="3rem" />
-      </UpArrow>
-
       <Carousel>
         {budgets.map((budget, index) => (
           <BudgetCard
@@ -54,9 +39,7 @@ const BudgetCardList: React.FC<IBudgetCardList> = ({ budgets, refetch }) => {
           />
         ))}
       </Carousel>
-      <DownArrow onClick={onDown}>
-        <IoIosArrowDown size="3rem" />
-      </DownArrow>
+
       <GotoCurrent onClick={() => setPosition(12)}>Goto Current</GotoCurrent>
     </Container>
   );
