+++ client/src/pages/Home/Home.styles.ts
@@ -5,15 +5,17 @@ export const Container = styled.div`
   position: relative;
   width: 100%;
   max-width: 800px;
-  height: calc(100vh - 70px);
+  height: calc(100vh - 90px);
   display: flex;
   align-items: center;
   margin: 0 auto;
+  justify-content: center;
+  user-select: none;
 `;
 
 export const PageTitle = styled.h3`
   position: absolute;
-  top: 0;
+  top: -10px;
   left: 50%;
   transform: translateX(-50%);
   text-align: center;
@@ -23,5 +25,29 @@ export const PageTitle = styled.h3`
 
   @media ${device.desktop} {
     font-size: 2rem;
+  }
+`;
+
+export const UpArrow = styled.div`
+  position: absolute;
+  top: 55px;
+  z-index: 1;
+  cursor: pointer;
+  transition: 0.3s all ease;
+
+  &:hover {
+    transform: scale(1.1);
+  }
+`;
+
+export const DownArrow = styled.div`
+  position: absolute;
+  bottom: 0;
+  z-index: 1;
+  cursor: pointer;
+  transition: 0.3s all ease;
+
+  &:hover {
+    transform: scale(1.1);
   }
 `;
